+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v 2.9.0 (?)
 Changes in this release:
+- Adds a new resource entity that raises an IgnoreResourceException when buildings its id
 
 # v 2.8.0 (2023-08-03)
 Changes in this release: